requirement.txt
streamlit>=1.28.0
requests>=2.31.0
google-generativeai>=0.3.0
plotly>=5.18.0